[trivial] fix time_limit of cegar_pattern and random_pattern generators

+++ src/search/pdbs/pattern_generator_cegar.h
@@ -10,7 +10,7 @@ class RandomNumberGenerator;
 namespace pdbs {
 class PatternGeneratorCEGAR : public PatternGenerator {
     const int max_pdb_size;
-    const int max_time;
+    const double max_time;
     const bool use_wildcard_plans;
     std::shared_ptr<utils::RandomNumberGenerator> rng;
