+++ src/search/pdbs/pattern_generator_random.h
@@ -10,7 +10,7 @@ class RandomNumberGenerator;
 namespace pdbs {
 class PatternGeneratorRandom : public PatternGenerator {
     const int max_pdb_size;
-    const int max_time;
+    const double max_time;
     const bool bidirectional;
     std::shared_ptr<utils::RandomNumberGenerator> rng;
